Handle warnings for the latest compilers.

+++ io/BpfWriter.cpp
@@ -262,7 +262,7 @@ void BpfWriter::writeView(const PointViewPtr dataShared)
     }
 
     size_t count = data->size() + m_header.m_numPts;
-    if (count > (std::numeric_limits<int32_t>::max)())
+    if (count > (size_t)(std::numeric_limits<int32_t>::max)())
         throwError("Too many points to write to BPF output. Limit is 2^32 -1.");
     m_header.m_numPts = static_cast<int32_t>(count);
 }
